requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.16
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
djangorestframework==3.15.2
Pillow==11.0.0
drf-yasg==1.21.8
uritemplate==4.1.1
inflection==0.5.1
pyyaml==6.0.2
pytz==2024.2
django-cors-headers==4.6.0
firebase-admin==6.6.0
google-cloud-storage==2.18.2
cachecontrol==0.14.1
pyjwt==2.10.0
google-cloud-firestore==2.19.0
google-api-python-client==2.154.0
google-api-core==2.23.0
msgpack==1.1.0
googleapis-common-protos==1.66.0
protobuf==5.28.3
google-auth==2.36.0
proto-plus==1.25.0
grpcio-status==1.68.0
grpcio==1.68.0
httplib2==0.22.0
google-auth-httplib2==0.2.0
pyasn1-modules==0.4.1
cachetools==5.5.0
rsa==4.9
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
pyparsing==3.2.0
pyasn1==0.6.1
cryptography==43.0.3
cffi==1.17.1
pycparser==2.22
